Update to use async and await

+++ server/server.js
@@ -67,28 +67,24 @@ app.get('/todos/:id', authenticate, (req, res) => {
 });
 
 // DELETE /todos/:id
-app.delete('/todos/:id', authenticate, (req, res) => {
-  // get the id
-  var id = req.params.id;
-  // validate the id -> not valid? return 404
+app.delete('/todos/:id', authenticate, async (req, res) => {
+  const id = req.params.id;
+  
   if (!ObjectID.isValid(id)) {
     return res.status(404).send();
   }
-  // remove todo by id
-  Todo.findOneAndRemove({
-    _id: id,
-    _creator: req.user._id
-  }).then((todo) => {
+  
+  try {
+    const todo = await Todo.findOneAndRemove({_id: id, _creator: req.user._id});
+  
     if (!todo) {
-      // if no doc, send 404
       return res.status(404).send();
     }
-    // if doc, send doc back with a 200
+    
     res.send({todo});
-  }).catch((e) => {
-    // 400 with empty body
-    res.status(400).send();
-  }); 
+  } catch (e) {
+    res.status(400).send();  
+  }
 });
 
 // PATCH /todos/:id
@@ -119,17 +115,16 @@ app.patch('/todos/:id', authenticate, (req, res) => {
 });
 
 // POST /users
-app.post('/users', (req, res) => {
-  var body = _.pick(req.body, ['email', 'password']);
-  var user = new User(body);
-
-  user.save().then(() => {
-    return user.generateAuthToken();
-  }).then((token) => {
+app.post('/users', async (req, res) => {
+  try {
+    const body = _.pick(req.body, ['email', 'password']);
+    const user = new User(body);
+    await user.save();
+    const token = await user.generateAuthToken();
     res.header('x-auth', token).send(user);
-  }).catch((e) => {
-    res.status(400).send(e);
-  });
+  } catch (e) {
+    res.status(400).send(e);  
+  }
 });
 
 // GET /users/me
@@ -138,25 +133,25 @@ app.get('/users/me', authenticate, (req, res) => {
 });
 
 // POST /users/login {email, password}
-app.post('/users/login', (req, res) => {
-  var body = _.pick(req.body, ['email', 'password']);
-
-  User.findByCredentials(body.email, body.password).then((user) => {
-    return user.generateAuthToken().then((token) => {
-      res.header('x-auth', token).send(user);
-    });
-  }).catch((e) => {
+app.post('/users/login', async (req, res) => {
+  try {
+    const body = _.pick(req.body, ['email', 'password']);
+    const user = await User.findByCredentials(body.email, body.password);
+    const token = await user.generateAuthToken();
+    res.header('x-auth', token).send(user);
+  } catch (e) {
     res.status(400).send();
-  });
+  }
 });
 
 // Private route
-app.delete('/users/me/token', authenticate, (req, res) => {
-  req.user.removeToken(req.token).then(() => {
+app.delete('/users/me/token', authenticate, async (req, res) => {
+  try {
+    await req.user.removeToken(req.token);
     res.status(200).send();
-  }, () => {
+  } catch (e) {
     res.status(400).send();
-  });
+  }
 });
 
 app.listen(port, () => {
